UPSTREAM: fscrypt: fix dereference of NULL user_key_payload
When an fscrypt-encrypted file is opened, we request the file's master
key from the keyrings service as a logon key, then access its payload.
However, a revoked key has a NULL payload, and we failed to check for
this.  request_key() *does* skip revoked keys, but there is still a
window where the key can be revoked before we acquire its semaphore.

Fix it by checking for a NULL payload, treating it like a key which was
already revoked at the time it was requested.

Fixes: 88bd6ccdcdd6 ("ext4 crypto: add encryption key management facilities")
Reviewed-by: James Morris <james.l.morris@oracle.com>
Cc: <stable@vger.kernel.org>    [v4.1+]
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: David Howells <dhowells@redhat.com>
(cherry-picked from commit d60b5b7854c3d135b869f74fb93eaf63cbb1991a
 and fixed up for android-3.18)
Change-Id: I7522324dc41969b5401fbe88ff90b4cd1f049cf2
Signed-off-by: Eric Biggers <ebiggers@google.com>

Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>